+++ packages/react-reconciler/src/ReactFiberBeginWork.js
