+++ packages/react-reconciler/src/ReactTypeOfSideEffect.js
